requirements.txt
Django==1.10
django-el-pagination==3.0
easy_thumbnails==2.3
postmarkup==1.2.2
BeautifulSoup==3.2.1
python-dateutil==2.5.3
django-constance==1.2
django_picklefield==0.3.2
django-bower==5.1.0
django-allauth==0.26
